removing filters from admin.py and creating a new module filters.py
