+++ basedosdados_api/api/v1/admin.py
@@ -7,6 +7,7 @@ from modeltranslation.admin import (
     TranslationStackedInline,
 )
 
+from basedosdados_api.api.v1.filters import OrganizationImageFilter, TableCoverageFilter
 from basedosdados_api.api.v1.models import (
     Organization,
     Dataset,
@@ -108,47 +109,6 @@ class TableInline(TranslationStackedInline):
 
 
 # Filters
-
-
-class OrganizationImageFilter(admin.SimpleListFilter):
-    title = "has_picture"
-    parameter_name = "has_picture"
-
-    def lookups(self, request, model_admin):
-        return (
-            ("True", "Yes"),
-            ("False", "No"),
-        )
-
-    def queryset(self, request, queryset):
-        if self.value() == "True":
-            return queryset.exclude(picture="")
-        if self.value() == "False":
-            return queryset.filter(picture="")
-
-
-class TableCoverageFilter(admin.SimpleListFilter):
-    title = "table_coverage"
-    parameter_name = "table_coverage"
-
-    def lookups(self, request, model_admin):
-        distinct_values = (
-            Coverage.objects.filter(table__id__isnull=False)
-            .order_by("area__name")
-            .distinct()
-            .values("area__name", "area__slug")
-        )
-        # Create a tuple of tuples with the format (value, label).
-        return [
-            (value.get("area__slug"), value.get("area__name"))
-            for value in distinct_values
-        ]
-
-        # return Coverage.objects.order_by().values("area__name").distinct()
-
-    def queryset(self, request, queryset):
-        if self.value():
-            return queryset.filter(coverages__area__slug=self.value())
 
 
 # Model Admins
@@ -253,15 +213,15 @@ class TableAdmin(TabbedTranslationAdmin):
             # If a parent model ID is provided, add the parent model field to the form
             # fields = self.get_related_fields
             initial = {"parent_model": parent_model_id}
-            self.initial = initial
+            self.initial = initial  # noqa
         return super().add_view(request, *args, **kwargs)
 
-    def get_related_fields(self, request, obj=None):
-        fields = self.model._meta.fields
+    def get_related_fields(self, request, obj=None):  # noqa
+        fields = self.model._meta.fields  # noqa
         parent_model_id = request.GET.get("dataset")
         if parent_model_id:
             parent_model = Dataset.objects.get(id=parent_model_id)
-            fields += parent_model._meta.fields
+            fields += parent_model._meta.fields  # noqa
         return fields
 
     readonly_fields = [
@@ -635,7 +595,7 @@ class QualityCheckAdmin(TabbedTranslationAdmin):
     ]
     search_fields = [
         "name",
-        "descriptiion",
+        "description",
     ]
     autocomplete_fields = [
         "analysis",
